handle notify signalR, send mail, fix some bugs

+++ green-craze-be-v1.Application/Model/Statistic/StatisticTopSellingProductYearResponse.cs
@@ -9,9 +9,9 @@ namespace green_craze_be_v1.Application.Model.Statistic
     public class StatisticTopSellingProductYearResponse
     {
         public string Date { get; set; }
-        public Dictionary<string, int> Products {  get; set; }
+        public Dictionary<string, long> Products { get; set; }
 
-        public StatisticTopSellingProductYearResponse(string date, Dictionary<string, int> products)
+        public StatisticTopSellingProductYearResponse(string date, Dictionary<string, long> products)
         {
             Date = date;
             Products = products;
